refactor: extracting cumulative hook
